Remove redundant cast

+++ starlark/library.go
@@ -665,7 +665,7 @@ func hash(thread *Thread, _ *Builtin, args Tuple, kwargs []Tuple) (Value, error)
 		return nil, fmt.Errorf("hash: got %s, want string or bytes", x.Type())
 	}
 	ret := MakeInt64(h)
-	if err := thread.AddAllocs(int64(EstimateSize(ret))); err != nil {
+	if err := thread.AddAllocs(EstimateSize(ret)); err != nil {
 		return nil, err
 	}
 	return ret, nil
